requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
azure-core==1.23.1
azure-functions==1.10.1
azure-functions-durable==1.1.3
azure-storage-blob==12.11.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
frozenlist==1.3.0
furl==2.1.3
idna==3.3
iniconfig==1.1.1
isodate==0.6.1
msrest==0.6.21
multidict==6.0.2
oauthlib==3.2.0
orderedmultidict==1.0.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyparsing==3.0.8
pytest==7.1.1
python-dateutil==2.8.2
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.9
yarl==1.7.2

# owned package
git+https://github.com/eduardo-prjadko/pypokemontcg.git